+++ Tests/QuillCodeParityTests/ParityTopBarGateTests.swift
@@ -34,16 +34,18 @@ final class ParityTopBarGateTests: QuillCodeParityTestCase {
     func testTopBarAgentStatusLabelsAreSharedByRuntimePaths() throws {
         let appStateText = try Self.appSourceText(named: "AppState.swift")
         let modelText = try Self.appSourceText(named: "WorkspaceModel.swift")
+        let reviewExtensionText = try Self.appSourceText(named: "WorkspaceModelReview.swift")
+        let workspaceRuntimeText = [modelText, reviewExtensionText].joined(separator: "\n")
         let builderText = try Self.appSourceText(named: "WorkspaceAgentStatusBuilder.swift")
         let mcpRuntimeText = try Self.appSourceText(named: "WorkspaceMCPRuntime.swift")
         let terminalLifecycleText = try Self.appSourceText(named: "WorkspaceTerminalLifecyclePlanner.swift")
 
         XCTAssertTrue(appStateText.contains("agentStatus: String = TopBarAgentStatusLabel.idle"), "Root state should use the shared idle label default.")
-        XCTAssertTrue(modelText.contains("TopBarAgentStatusLabel.running"), "WorkspaceModel should use shared running status copy.")
+        XCTAssertTrue(workspaceRuntimeText.contains("TopBarAgentStatusLabel.running"), "Workspace runtime paths should use shared running status copy.")
         XCTAssertTrue(terminalLifecycleText.contains("TopBarAgentStatusLabel.terminal"), "Terminal lifecycle planner should use shared terminal status copy.")
         XCTAssertTrue(builderText.contains("TopBarAgentStatusLabel.streaming"), "Agent progress builder should use shared streaming status copy.")
         XCTAssertTrue(mcpRuntimeText.contains("TopBarAgentStatusLabel.failed"), "MCP runtime should use shared failed status copy.")
-        XCTAssertFalse(modelText.contains("refreshTopBar(agentStatus: \""), "WorkspaceModel should not pass raw lifecycle status strings to the top bar.")
+        XCTAssertFalse(workspaceRuntimeText.contains("refreshTopBar(agentStatus: \""), "Workspace runtime paths should not pass raw lifecycle status strings to the top bar.")
         XCTAssertFalse(builderText.contains("return \"Running\""), "Agent progress builder should not return raw lifecycle status strings.")
         XCTAssertFalse(builderText.contains("return \"Failed\""), "Agent progress builder should not return raw lifecycle status strings.")
         XCTAssertFalse(mcpRuntimeText.contains("agentStatus: \"Idle\""), "MCP runtime should not return raw idle status strings.")
